Statics in Stripe must be volatile
According to the [Java Memory Model](http://docs.oracle.com/javase/specs/jls/se7/html/jls-17.html#jls-17.4)
writes to a shared variable from one thread are only visible to an
other thread if there is a happens-before relationship between them.
There currently is no happens-before relationship between the writes
to `Stripe` `apiKey`, `apiVersion`, `apiBase` and the reads in
`APIResource`. There are several ways to introduce a happens-before
relationship between the writes and the reads but the easiest is to
mark the fields volatile.