Adding support for posting multipart/form data and file uploads.
This commit does a number of things in order to allow file uploads to be
included in the java bindings. These are:

1. It allows subclasses of APIResource to specify apiBase when
constructing a URL. This allows a method to point to base endpoints
different than `https://api.stripe.com`.

2. The multipartRequest method in APIResource allows one to post
multipart/form-data. To send a file, one simply needs to pass a
java.io.File as a value in params.

3. Creates the FileUpload resource which can be used to create and
retrieve file uploads from the Stripe upload API.

We're also bumping the version since this introduces changes into the
APIResource class.