Manually enable TLSv1.2 for the benefit of Java 6 and 7
Java 6 and 7 both support TLSv1.2, but don't enable it (or TLSv1.1) by
default for TLS connections due to concerns at release-time around
server compatibility. At this point, those are largely addressed, and
TLSv1.0 is being deprecated by, among others, the PCI DSS within the
next year.

We don't want to pin specifically to TLSv1.2, as that makes us not
forward-compatible when TLSv1.3 (and later) is eventually
released. However, the only way I could find to update the list of
protocol versions was to call the setEnabledProtocols method on an
SSLSocket.

And the only way I could find to do that was to provide my own
SSLSocketFactory to HttpsURLConnection, which this does. The new
factory accepts an "underlying" factory as an argument, and proxies
all calls to it, enabling TLSv1.2 on the resulting sockets as
appropriate.