Detect support for TLSv1.1 and TLSv1.2 and only enable when supported
Sufficiently old JREs might not support TLSv1.1 or 1.2 at all, which
is pretty unfortunate, but we should probably do better than just
crashing with an `IllegalArgumentException`.

In local testing, this also seems to fix the mysterious "Not enough
data to fill declared vector size" exception.  At a guess, enabling
TLSv1.2 but not TLSv1.1 confuses the handshake code.