Fix encoding of Lists being sent in via API requests
As described in #229, we were previously translating lists to a special
type of indexed object when sending them into the API via a request.
This patch corrects that behavior, adds some significant clean-up to the
encoding utilities, and adds some additional testing so that we'll have
better guarantees that everything works.