Fix non-deterministic test
Something has changed recently (probably in the server-side
implementation on which these bindings rely) where a specific sorting
of order items coming back in an order is no longer guaranteed. These
tests were written so that in a list of (1) sku, (2) shipping, and (3)
tax, it was always assumed that the sku would appear first. This
assumption no longer hold ups.

Patch now searches order items for sku and compares against that instead
of assuming that it's in 0th position.