Elaborate on paging iterator test to demonstrate failure
This extends the paging iterator test from two pages to three in order
to get some test coverage on the sections where we pull a page from a
collection that is *not* the original collection. Not having this piece
is what allowed the bug in #237 to surface despite having some testing
in place.

This should be rebased and brought in after the fix for #237 in #239 has
been brought in.