Allow params/options to be passed when paging through subcollections
Currently when a list is fetched from the API params and options can be
given to it as long as it's for a top-level type. If a resource includes
a subcollection (e.g. line items under invoices), the current page can
be iterated, but future pages that are fetched do not inherit the params
and options that were given to the parent.

This change suggests a breaking change to any subcollection getters
(e.g. getLines) so that we can return proxy objects which will allow use
a little like the following:

``` java
items = invoice.getLines().autoPagingIterable(params, options)
```

It would also be nice to do the same for top-level resources so that
this:

``` java
invoices = Invoice.list(params, options).autoPagingIterable()
```

Might become this:

``` java
invoices = Invoice.list().autoPagingIterable(params, options)
```

For consistency's sake. This turns out be a little more of a yakshave in
practice though because we need to have the proxy class remember a
collection type as well as the resource's type.

Fixes #264.