Allow request params and options to be overridden on `autoPagingIterator`
We normally pass params and options with the `list` method:

    invoices = Invoice.list(params, options).autoPagingIterable()

This patch adds support for them to also be overridden from
`autoPagingIterator`:

    invoices = Invoice.list().autoPagingIterable(params, options)

The reason for this is to support custom params and options for
subcollections, which unfortunately do not inherit options from their
parent collections. For example:

    lineItems = invoice.getLines().autoPagingIterable(params, options)

There's more context for this in #264, and I tried a more elegant
solution (which failed) in #269. This seems like the easiest option for
the time being, with the possibility that we could do something more
sophisticated later to start consolidating our interfaces.

Fixes #264.