Remove StripeAPIResourceCollection
The current implementation has two collection implementations:

* `StripeAPIResourceCollection`: Extends `APIResource`; used when a
  collection needs access to either `request` or `requestCollection`.
* `StripeAPICollection`: Used in all other situations.

The two collection implementations are functionally identical except for
their parent class. This is somewhat error prone because as seen in #292,
it's easy to add something to one and forget about the other.

This pull eliminates `StripeAPIResourceCollection` and moves all
collections over to just extend `StripeCollection`, solving the
duplication problem and reducing code complexity.

The downside is that I've moved visibility of `request` and
`requestCollection` on `APIResource` to public so that they can be
called into from collection classes. It would be better if these had
something like `internal` visibility (a la C#), but Java doesn't really
support anything like that. IMO, it's still a reasonable trade-off and
we should still consider those `request*` methods "effectively" private
to the package and not part of the public API.

Fixes #292.