Add test to demonstrate event deserialization
Adds a basic test to demonstrate how to deserialize an event and then
typecast the `StripeObject` nested in it to a model that's appropriate
based on the type of event.

Not a true fix for #297, but at least produces some sample code for this
operation, even if not particularly discoverable.