Outcomes on charges are now collapsed by default
This makes outcomes on charges an expandable field. We leave the old
`getOutcome` and `setOutcome` in place to try and maintain backwards
compatibility.

To compensate for the reserved names, I introduce `getOutcomeID` and
`setOutcomeID` instead, which aren't really a convention right now (but
maybe should be?).

We should also consider just breaking thd old version of the field.
That's up for discussion.