Make `rule` under charge outcome expandable
Adds accessors for an expandable `rule` under charge outcome now that
this field is now collapsed by default. We also start assigning
`ChargeOutcomeRule` and ID with `HasId`.

I changed the tests around a little bit to make them a little more
consistent with how we test expansions elsewhere in the suite.

Fixes #355.