Merge expand tests into one and no more list
This list expansion test was basically selecting a charge at random from
the test account and hoping it had a balance transaction when it might
not. This was especially problematic when other suites were running
concurrently and creating charges which might still be uncaptured.

We really don't need to check expansions on list anyway (and probably
not for both create and charge either). If it works in any of these
places, it's going to work for all of them.