Add event request object
Event's `request` field has changed from being a string to being an
object that contains a request ID and an idempotency key.

I don't think there's an easy way of making this backwards compatible,
and since it's probably not in all that common of use, I've just changed
the type completely here.