Power test suite with external API stub
Move to using an external API stub in the test suite instead of issuing live API calls so that we can have tests that run more quickly and more reliably.

For the time being, I've just converted the plan test suite over to use the new scheme.

This is an experimental prototype for demonstrative purposes and should not be merged.

See also:

* https://github.com/stripe/stripe-go/issues/424
* https://github.com/stripe/stripe-php/pull/357
* https://github.com/stripe/stripe-ruby/pull/553