Only set value on DNS cache TTL if it's been modified
While making a request, stripe-java sets the DNS cache's TTL to zero. I
traced the commit back to 01fd2dba2ec1885f68c52355813eece3975ba3f1, but
unfortunately the author didn't leave us a shred of information as to
why they did this. I'd guess that it was was done because people were
setting long cache times and had originally resolved the host
incorrectly. I don't think it's related to security, because your TLS
certificate validation will still be active even if an attacker managed
to poison the DNS cache.

Here we take svilenivanovsmule's suggestion in #405 and only modify the
cache TTL *if* the user has overridden it. Java has pretty reasonable
defaults for this, and we shouldn't really need to touch it (unless
someone can think of a good reason why we would).

I think another very plausible option here is to just strip out all of
this code completely -- even someone who's written Java for only about
four minutes should've been able to tell that this code is not even
remotely thread-safe as written. Given that fact and no known reason why
this should be here, taking it out might be the right way to go.

Fixes #405.