Implement InputStream based FileUpload creation
This commit adds in suppor for calling
FileUpload.create(Map<String, Object>) with the parameter "file" set to
an InputStream, instead of only allowing Files. You can see an example
of this in src/test/java/com/stripe/functional/FileUploadTest.java.

Furthermore, this commit changes the addFileField method to no longer
accept a file argument, and only allows for InputStream arguments in
addition to a string containing the name of the file to upload.