Fix retrieval of upcoming Invoice line items
Some endpoints receive the "root" object id as a query string parameter for example

`/v1/invoices/upcoming/lines?customer=cus_xxx`

if `invoice.getLines().list(linesParams)` is used. If, in `linesParams`, further
options are passed (limit: 5 for example), these are incorrectly appended to
the URI using a question mark and not an ampersand, resulting in calls
for upcoming line items for a customer specified by a resource ID of cus_xxx?limit=5.