Introduce generated example functional test
In addition to the API spec used to generate most of this SDK, Stripe also
maintains a corpus of example requests that are expected to work consistently
across all supported languages.

This introduces a new functional test, generated from the spec and example
corpus, to increase confidence that critical flows and use-cases work as
expected in Java.